+++ map/index.cjs
@@ -1,0 +1,2 @@
+exports = module.exports = require('../commonjs/read/schema/convertToJson.js').default
+exports['default'] = require('../commonjs/read/schema/convertToJson.js').default