+++ map/index.js
@@ -1,0 +1,1 @@
+export { default as default } from '../modules/read/schema/convertToJson.js'