+++ schema/index.cjs
@@ -1,2 +1,2 @@
-exports = module.exports = require('../commonjs/read/schema/convertToJson.js').default
-exports['default'] = require('../commonjs/read/schema/convertToJson.js').default
+exports = module.exports = require('../commonjs/read/schema/convertToJson.legacy.js').default
+exports['default'] = require('../commonjs/read/schema/convertToJson.legacy.js').default