+++ schema/index.d.ts
@@ -1,6 +1,11 @@
 import {
 	Row,
 	Schema
-} from '../types.d';
+} from '../types.d.js';
 
-export default function convert(data: Row[], schema: Schema): object[];
+export default function mapWithLegacyBehavior(data: Row[], schema: Schema, options?: {
+	ignoreEmptyRows?: boolean,
+	includeNullValues?: boolean,
+	isColumnOriented?: boolean,
+	rowMap?: Record<string, number>
+}): object[];