+++ source/read/schema/convertToJson.js
@@ -4,15 +4,33 @@ import BooleanType from '../../types/Boolean.js'
 import DateType from '../../types/Date.js'
 
 const DEFAULT_OPTIONS = {
-  isColumnOriented: false
+  schemaPropertyValueForMissingColumn: undefined,
+  schemaPropertyValueForUndefinedCellValue: undefined,
+  schemaPropertyValueForNullCellValue: null,
+  schemaPropertyShouldSkipRequiredValidationForMissingColumn: () => false,
+  // `getEmptyObjectValue(object, { path })` applies to both the top-level object
+  // and any of its sub-objects.
+  getEmptyObjectValue: () => null,
+  getEmptyArrayValue: () => null,
+  isColumnOriented: false,
+  arrayValueSeparator: ','
 }
 
 /**
- * Convert 2D array to nested objects.
- * If row oriented data, row 0 is dotted key names.
- * Column oriented data is transposed.
+ * (this function is exported from `read-excel-file/map`)
+ * Converts spreadsheet-alike data structure into an array of objects.
+ * The first row should be the list of column headers.
  * @param {any[][]} data - An array of rows, each row being an array of cells.
  * @param {object} schema
+ * @param {object} [options]
+ * @param {null} [options.schemaPropertyValueForMissingColumn] — By default, when some of the `schema` columns are missing in the input `data`, those properties are set to `undefined` in the output objects. Pass `schemaPropertyValueForMissingColumn: null` to set such "missing column" properties to `null` in the output objects.
+ * @param {null} [options.schemaPropertyValueForNullCellValue] — By default, when it encounters a `null` value in a cell in input `data`, it sets it to `undefined` in the output object. Pass `schemaPropertyValueForNullCellValue: null` to make it set such values as `null`s in output objects.
+ * @param {null} [options.schemaPropertyValueForUndefinedCellValue] — By default, when it encounters an `undefined` value in a cell in input `data`, it it sets it to `undefined` in the output object. Pass `schemaPropertyValueForUndefinedCellValue: null` to make it set such values as `null`s in output objects.
+ * @param {boolean} [options.schemaPropertyShouldSkipRequiredValidationForMissingColumn(column: string, { object })] — By default, it does apply `required` validation to `schema` properties for which columns are missing in the input `data`. One could pass a custom `schemaPropertyShouldSkipRequiredValidationForMissingColumn(column, { object })` to disable `required` validation for missing columns in some or all cases.
+ * @param {function} [options.getEmptyObjectValue(object, { path })] — By default, it returns `null` for an "empty" resulting object. One could override that value using `getEmptyObjectValue(object, { path })` parameter. The value applies to both top-level object and any nested sub-objects in case of a nested schema, hence the additional `path?: string` parameter.
+ * @param {function} [getEmptyArrayValue(array, { path })] — By default, it returns `null` for an "empty" array value. One could override that value using `getEmptyArrayValue(array, { path })` parameter.
+ * @param {boolean} [options.isColumnOriented] — By default, the headers are assumed to be the first row in the `data`. Pass `isColumnOriented: true` if the headers are the first column in the `data`. i.e. if `data` is "transposed".
+ * @param {object} [options.rowIndexMap] — Custom row index mapping `data` rows. If present, will overwrite the indexes of `data` rows with the indexes from this `rowIndexMap`.
  * @return {object[]}
  */
 export default function(data, schema, options) {
@@ -27,8 +45,7 @@ export default function(data, schema, options) {
 
   const {
     isColumnOriented,
-    rowMap,
-    ignoreEmptyRows
+    rowIndexMap
   } = options
 
   validateSchema(schema)
@@ -43,20 +60,18 @@ export default function(data, schema, options) {
   const errors = []
 
   for (let i = 1; i < data.length; i++) {
-    const result = read(schema, data[i], i, columns, errors, options)
-    if (result !== null || ignoreEmptyRows === false) {
-      results.push(result)
-    }
+    const result = read(schema, data[i], i, undefined, columns, errors, options)
+    results.push(result)
   }
 
-  // Correct error rows.
-  if (rowMap) {
+  // Set the correct `row` number in `errors` if a custom `rowIndexMap` is supplied.
+  if (rowIndexMap) {
     for (const error of errors) {
       // Convert the `row` index in `data` to the
       // actual `row` index in the spreadsheet.
       // `- 1` converts row number to row index.
       // `+ 1` converts row index to row number.
-      error.row = rowMap[error.row - 1] + 1
+      error.row = rowIndexMap[error.row - 1] + 1
     }
   }
 
@@ -66,7 +81,7 @@ export default function(data, schema, options) {
   }
 }
 
-function read(schema, row, rowIndex, columns, errors, options) {
+function read(schema, row, rowIndex, path, columns, errors, options) {
   const object = {}
   let isEmptyObject = true
 
@@ -93,59 +108,78 @@ function read(schema, row, rowIndex, columns, errors, options) {
 
   const pendingRequiredChecks = []
 
+  // For each schema entry.
   for (const key of Object.keys(schema)) {
     const schemaEntry = schema[key]
     const isNestedSchema = typeof schemaEntry.type === 'object' && !Array.isArray(schemaEntry.type)
 
-    let rawValue = row[columns.indexOf(key)]
-    if (rawValue === undefined) {
-      rawValue = null
+    // The path of this property inside the resulting object.
+    const propertyPath = `${path || ''}.${schemaEntry.prop}`
+
+    // Read the cell value for the schema entry.
+    let cellValue
+    const columnIndex = columns.indexOf(key)
+    const isMissingColumn = columnIndex < 0
+    if (!isMissingColumn) {
+      cellValue = row[columnIndex]
     }
 
     let value
     let error
     let reason
 
+    // Get property `value` from cell value.
     if (isNestedSchema) {
-      value = read(schemaEntry.type, row, rowIndex, columns, errors, options)
+      value = read(schemaEntry.type, row, rowIndex, propertyPath, columns, errors, options)
     } else {
-      if (rawValue === null) {
-        value = null
+      if (isMissingColumn) {
+        value = options.schemaPropertyValueForMissingColumn
+      }
+      else if (cellValue === undefined) {
+        value = options.schemaPropertyValueForUndefinedCellValue
+      }
+      else if (cellValue === null) {
+        value = options.schemaPropertyValueForNullCellValue
       }
       else if (Array.isArray(schemaEntry.type)) {
-        let notEmpty = false
-        const array = parseArray(rawValue).map((_value) => {
+        const array = parseArray(cellValue, options.arrayValueSeparator).map((_value) => {
+          if (error) {
+            return
+          }
           const result = parseValue(_value, schemaEntry, options)
           if (result.error) {
+            // In case of an error, `value` won't be returned and will just be reported
+            // as part of an `error` object, so it's fine assigning just an element of the array.
             value = _value
             error = result.error
             reason = result.reason
           }
-          if (result.value !== null) {
-            notEmpty = true
-          }
           return result.value
         })
         if (!error) {
-          value = notEmpty ? array : null
+          const isEmpty = array.every(isEmptyValue)
+          value = isEmpty ? options.getEmptyArrayValue(array, { path: propertyPath }) : array
         }
       } else {
-        const result = parseValue(rawValue, schemaEntry, options)
+        const result = parseValue(cellValue, schemaEntry, options)
         error = result.error
         reason = result.reason
-        value = error ? rawValue : result.value
+        value = error ? cellValue : result.value
       }
     }
 
-    if (!error && value === null) {
-      if (typeof schemaEntry.required === 'function') {
-        pendingRequiredChecks.push({ column: key })
-      } else if (schemaEntry.required === true) {
-        error = 'required'
+    // Apply `required` validation if the value is "empty".
+    if (!error && isEmptyValue(value)) {
+      if (schemaEntry.required) {
+        // Will perform this `required()` validation in the end,
+        // when all properties of the mapped object have been mapped.
+        pendingRequiredChecks.push({ column: key, value, isMissingColumn })
       }
     }
 
     if (error) {
+      // If there was an error then the property value in the `object` will be `undefined`,
+      // i.e it won't add the property value to the mapped object.
       errors.push(createError({
         column: key,
         value,
@@ -153,30 +187,41 @@ function read(schema, row, rowIndex, columns, errors, options) {
         reason
       }))
     } else {
-      if (isEmptyObject && value !== null) {
+      // Possibly unmark the mapped object as "empty".
+      if (isEmptyObject && !isEmptyValue(value)) {
         isEmptyObject = false
       }
-      if (value !== null || options.includeNullValues) {
+      // Set the value in the mapped object.
+      // Skip setting `undefined` values because they're already `undefined`.
+      if (value !== undefined) {
         object[schemaEntry.prop] = value
       }
     }
   }
 
+  // Return `null` for an "empty" mapped object.
   if (isEmptyObject) {
-    return null
+    return options.getEmptyObjectValue(object, { path })
   }
 
-  for (const { column } of pendingRequiredChecks) {
-    const required = schema[column].required(object)
-    if (required) {
-      errors.push(createError({
-        column,
-        value: null,
-        error: 'required'
-      }))
+  // Perform any `required` validations.
+  for (const { column, value, isMissingColumn } of pendingRequiredChecks) {
+    // Can optionally skip `required` validation for missing columns.
+    const skipRequiredValidation = isMissingColumn && options.schemaPropertyShouldSkipRequiredValidationForMissingColumn(column, { object })
+    if (!skipRequiredValidation) {
+      const { required } = schema[column]
+      const isRequired = typeof required === 'boolean' ? required : required(object)
+      if (isRequired) {
+        errors.push(createError({
+          column,
+          value,
+          error: 'required'
+        }))
+      }
     }
   }
 
+  // Return the mapped object.
   return object
 }
 
@@ -305,12 +350,12 @@ export function getBlock(string, endCharacter, startIndex) {
  * @param  {string} string — A string of comma-separated substrings.
  * @return {string[]} An array of substrings.
  */
-export function parseArray(string) {
+export function parseArray(string, arrayValueSeparator) {
   const blocks = []
   let index = 0
   while (index < string.length) {
-    const [substring, length] = getBlock(string, ',', index)
-    index += length + ','.length
+    const [substring, length] = getBlock(string, arrayValueSeparator, index)
+    index += length + arrayValueSeparator.length
     blocks.push(substring.trim())
   }
   return blocks
@@ -327,4 +372,8 @@ function validateSchema(schema) {
       throw new Error(`"prop" not defined for schema entry "${key}".`)
     }
   }
+}
+
+function isEmptyValue(value) {
+  return value === undefined || value === null
 }